Changes to support overriding surveyor helper methods (in same way as models are overridden) Issue 213
