+++ lib/generators/surveyor/templates/surveys/EXTENDING_SURVEYOR.md
@@ -29,6 +29,20 @@ For example, create app/models/survey.rb with the following contents:
 
 == SurveyorHelper
 
+The SurveyorHelper module can be customized just like the models:
+
+    module SurveyorHelper
+      include Surveyor::Helpers::SurveyorHelperMethods
+      def rc_to_as(type_sym)
+        case type_sym.to_s
+        when /(integer|float)/ then :string
+        when /(datetime)/ then :string
+        when /^date$/ then :string
+        else type_sym
+        end
+      end
+    end
+
 == Views
 
 Surveyor's views can be overwritten by simply creating views in app/views/surveyor
