+++ spec/helpers/surveyor_helper_spec.rb
@@ -1,8 +1,9 @@
 require File.expand_path(File.dirname(__FILE__) + '/../spec_helper')
 
 describe SurveyorHelper do
+
   before(:each) do
-    
+
   end
   it "should return the question text with number" do
     q1 = Factory(:question)
@@ -13,12 +14,12 @@ describe SurveyorHelper do
     helper.q_text(q1).should == "1) #{q1.text}"
     helper.q_text(q2).should == q2.text
     helper.q_text(q3).should == q3.text
-    
+
     dir = "images"
     if Rails.application.config.respond_to?(:assets) && Rails.application.config.assets
       dir = "assets"
     end
-    
+
     helper.q_text(q4).should == %Q(<img alt="Something" src="/#{dir}/something.jpg" />)
     helper.q_text(q5).should == q5.text
   end
@@ -30,7 +31,7 @@ describe SurveyorHelper do
     def somethingElse
       "something new"
     end
-    
+
     def group
       "NUBIC"
     end
@@ -42,7 +43,7 @@ describe SurveyorHelper do
     helper.q_text(label, FakeMustacheContext).should == "Testing something new"
   end
   it "should return help_text for question with substituted value" do
-    q2 = Factory(:question, :display_type => "label", :text => "Is you site Northwestern?", :help_text => "If your site is not {{site}}, pick 'no' for the answer") 
+    q2 = Factory(:question, :display_type => "label", :text => "Is you site Northwestern?", :help_text => "If your site is not {{site}}, pick 'no' for the answer")
     helper.render_help_text(q2, FakeMustacheContext).should == "If your site is not Northwestern, pick 'no' for the answer"
   end
   it "should return help_text for group text with number" do
@@ -57,9 +58,9 @@ describe SurveyorHelper do
   it "should return rendered text for answer" do
     q1 = Factory(:question, :text => "Do you work for {{site}}", :answers => [a1 = Factory(:answer, :text => "No, I don't work for {{site}}"), a2 = Factory(:answer, :text => "Yes, I do work for {{site}}") ])
     helper.a_text(a1, nil, FakeMustacheContext).should == "No, I don't work for Northwestern"
-    helper.a_text(a2, nil, FakeMustacheContext).should == "Yes, I do work for Northwestern"    
+    helper.a_text(a2, nil, FakeMustacheContext).should == "Yes, I do work for Northwestern"
   end
-  
+
   it "should return the group text with number" do
     g1 = Factory(:question_group)
     helper.q_text(g1, FakeMustacheContext).should == "1) #{g1.text}"
@@ -78,7 +79,7 @@ describe SurveyorHelper do
     helper.response_for(rs, q2).attributes.should == Response.new(:question => q2, :response_set => rs).attributes
     helper.response_for(rs, q2, b).attributes.should == Response.new(:question => q2, :response_set => rs).attributes
     helper.response_for(rs, q3, c, "1").should == r3
-    
+
   end
   it "should keep an index of responses" do
     helper.response_idx.should == "1"
@@ -95,4 +96,41 @@ describe SurveyorHelper do
     helper.rc_to_attr(:date).should == :datetime_value
     helper.rc_to_attr(:time).should == :datetime_value
   end
+
+  it "should translate response class into as" do
+    helper.rc_to_as(:string).should == :string
+    helper.rc_to_as(:text).should == :text
+    helper.rc_to_as(:integer).should == :string
+    helper.rc_to_as(:float).should == :string
+    helper.rc_to_as(:datetime).should == :datetime
+    helper.rc_to_as(:date).should == :string
+    helper.rc_to_as(:time).should == :time
+  end
+
+  context "overriding methods" do
+    before do
+      module SurveyorHelper
+        include Surveyor::Helpers::SurveyorHelperMethods
+
+        def rc_to_as(type_sym)
+          case type_sym.to_s
+          when /(integer|float)/ then :string
+          when /(datetime)/ then :string
+          else type_sym
+          end
+        end
+      end
+    end
+
+    it "should translate response class into as" do
+      helper.rc_to_as(:string).should == :string
+      helper.rc_to_as(:text).should == :text
+      helper.rc_to_as(:integer).should == :string
+      helper.rc_to_as(:float).should == :string
+      helper.rc_to_as(:datetime).should == :string  # not datetime
+      helper.rc_to_as(:date).should == :date        # not string
+      helper.rc_to_as(:time).should == :time
+    end
+  end
+
 end
